pyproject.toml
[project]
name = "rl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datasets==3.6.0",
    "future>=1.0.0",
    "google-generativeai>=0.8.5",
    "gym>=0.26.2",
    "librosa>=0.11.0",
    "matplotlib>=3.10.3",
    "numpy>=2.2.6",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.2",
    "pypdf>=6.1.0",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.7.1",
    "setuptools>=80.9.0",
    "six>=1.17.0",
    "soundfile>=0.13.1",
    "tinygrad>=0.10.3",
    "tokenizers>=0.21.4",
    "torch>=2.7.1",
    "torchcodec>=0.7.0",
    "torchvision>=0.22.1",
    "tqdm>=4.67.1",
    "transformers>=4.55.2",
    "typing>=3.10.0.0",
]
